+++ src/core/processor/mod.rs
@@ -19,6 +19,7 @@ use crate::{
 };
 
 mod helpers;
+
 pub mod read;
 pub mod write;